changed some data types to reduce memory required

+++ EightBitTunes/TuneManager.h
@@ -6,20 +6,20 @@
 
 // HARDWARE CONSTANTS
 // pin of speaker/pezo
-static const int PIN_PEZO = 8;
-static const int PIN_CS = 4;
-static const int PIN_HARDWARE_SS = 10;
+static const byte PIN_PEZO = 8;
+static const byte PIN_CS = 4;
+static const byte PIN_HARDWARE_SS = 10;
 
 // Buffer Constants
-static const int MIN_NOTE_BUFFER = 15;
-static const int MAX_NOTE_BUFFER = 100;
+static const byte MIN_NOTE_BUFFER = 15;
+static const byte MAX_NOTE_BUFFER = 100;
 
 class TuneManager {
   public:
     TuneManager(char tuneFolderPath[]); 
     void playTunes();
   private:
-    void addNotesToTune(Stream* str, int numOfNotesToAdd);
+    void addNotesToTune(Stream* str, byte numOfNotesToAdd);
     void ensureFile(File* file, File* rootToLoadFrom);
 };
